requirements.txt
attrs==22.1.0
bandit==1.7.4
black==22.10.0
click==8.1.3
coverage==6.5.0
exceptiongroup==1.0.4
flake8==6.0.0
gitdb==4.0.10
GitPython==3.1.29
grpcio==1.50.0
grpcio-reflection==1.50.0
grpcio-tools==1.50.0
iniconfig==1.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
pbr==5.11.0
platformdirs==2.5.4
pluggy==1.0.0
protobuf==4.21.9
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
PyYAML==6.0
six==1.16.0
smmap==5.0.0
stevedore==4.1.1
tomli==2.0.1
